DOS 2/3

+++ Assets/Script/Request_Data.cs
@@ -17,7 +17,7 @@ public class Day_data : Month_Data
 
 public class Check_daily_request : Request_Data{
     public string[] waktu;
-    public int[] total_request,total_object,total_client;
+    public long[] total_request,total_object,total_client;
 
 }
 public class Check_daily_request_by_month : Month_Data{
@@ -26,7 +26,7 @@ public class Check_daily_request_by_month : Month_Data{
 }
 public class Check_request_on_day : Day_data{
     public string[] ip_address;
-    public int[] total_request,total_object;
+    public long[] total_request,total_object;
 }
 public class Get_request_line_and_status_code : Day_data{
     public string ip_address;
@@ -37,25 +37,25 @@ public class Get_request_line_and_status_code : Day_data{
 public class Search_usage_of_keyword : Request_Data{
     public string keyword;
     public string[] ip_address;
-    public int[] total_request;
+    public long[] total_request;
 }
 public class Search_usage_of_keyword_on_month : Month_Data{
     public string keyword;
     public string[] ip_address;
-    public int[] total_request;
+    public long[] total_request;
 }
 public class Check_status_code_occurence : Request_Data{
     public string keyword;
     public string[] waktu;
-    public int[] total_status;
+    public long[] total_status;
 }
 public class Check_status_code_occurence_on_month : Month_Data{
     public string keyword;
     public string[] waktu;
-    public int[] total_status;
+    public long[] total_status;
 }
 public class Check_status_code_occurence_per_ip : Day_data{
     public string keyword;
     public string[] ip_address;
-    public int[] total_status;
+    public long[] total_status;
 }
